+++ modules/Accounting/includes/DailyTransactions.php
@@ -12,6 +12,16 @@ $start_date = RequestedDate( 'start', date( 'Y-m' ) . '-01' );
 // Set end date.
 $end_date = RequestedDate( 'end', DBDate() );
 
+$_REQUEST['category'] = issetVal( $_REQUEST['category'], '' );
+
+// Sanitize category (type or ID).
+if ( ! empty( $_REQUEST['category'] )
+	&& ! in_array( $_REQUEST['category'], [ 'common', 'incomes', 'expenses'] )
+	&& (string) (int) $_REQUEST['category'] !== $_REQUEST['category'] )
+{
+	$_REQUEST['category'] = '';
+}
+
 if ( User( 'PROFILE' ) === 'admin' )
 {
 	DrawHeader( _programMenu( 'transactions' ) );
@@ -22,15 +32,17 @@ echo '<form action="' . URLEscape( 'Modules.php?modname=' . $_REQUEST['modname']
 $header_checkboxes = '<label><input type="checkbox" value="true" name="accounting" id="accounting" ' .
 ( ! isset( $_REQUEST['accounting'] )
 	|| $_REQUEST['accounting'] == 'true' ? 'checked ' : '' ) . '/> ' .
-_( 'Income' ) . ' & ' . _( 'Expense' ) . '</label>&nbsp; ';
+_( 'Income' ) . ' & ' . _( 'Expense' ) . '</label>';
 
-$header_checkboxes .= '<label><input type="checkbox" value="true" name="staff_payroll" id="staff_payroll" ' .
+$header_checkboxes .= ' &mdash; <label>' . _( 'Category' ) . ' ' . _categorySelect( $_REQUEST['category'] ) . '</label> ';
+
+$header_checkboxes .= '<br /><label><input type="checkbox" value="true" name="staff_payroll" id="staff_payroll" ' .
 ( ! empty( $_REQUEST['staff_payroll'] ) ? 'checked ' : '' ) . '/> ' .
-_( 'Staff Payroll' ) . '</label>&nbsp; ';
+_( 'Staff Payroll' ) . '</label>';
 
 if ( $RosarioModules['Student_Billing'] )
 {
-	$header_checkboxes .= '<label><input type="checkbox" value="true" name="student_billing" id="student_billing" ' .
+	$header_checkboxes .= '<br /><label><input type="checkbox" value="true" name="student_billing" id="student_billing" ' .
 	( ! empty( $_REQUEST['student_billing'] ) ? 'checked ' : '' ) . '/> ' .
 	_( 'Student Billing' ) . '</label>';
 }
@@ -69,20 +81,40 @@ if ( ! isset( $_REQUEST['accounting'] )
 		$name_col_sql = "'' AS FULL_NAME,";
 	}
 
+	$where_category_sql = '';
+
+	if ( is_numeric( $_REQUEST['category'] ) )
+	{
+		$where_category_sql = " AND CATEGORY_ID='" . (int) $_REQUEST['category'] . "'";
+
+		if ( ! $_REQUEST['category'] )
+		{
+			// "0", N/A => Category ID is NULL.
+			$where_category_sql = " AND CATEGORY_ID IS NULL";
+		}
+	}
+	elseif ( $_REQUEST['category'] )
+	{
+		$where_category_sql = " AND CATEGORY_ID IN (SELECT ID
+			FROM accounting_categories
+			WHERE SCHOOL_ID='" . Userschool() . "'
+			AND TYPE='" . $_REQUEST['category'] . "')";
+	}
+
 	$RET = DBGet( "SELECT " . $name_col_sql . "f.AMOUNT AS CREDIT,'' AS DEBIT,CONCAT(f.TITLE,' ',COALESCE(f.COMMENTS,'')) AS EXPLANATION,f.ASSIGNED_DATE AS DATE,f.ID AS ID
 	FROM accounting_incomes f
 	WHERE f.SYEAR='" . UserSyear() . "'
 	AND f.SCHOOL_ID='" . UserSchool() . "'
 	AND f.ASSIGNED_DATE BETWEEN '" . $start_date . "'
-	AND '" . $end_date . "'", $extra['functions'] );
+	AND '" . $end_date . "'" . $where_category_sql, $extra['functions'] );
 
-	$payments_SQL = "SELECT " . $name_col_sql . "'' AS CREDIT,p.AMOUNT AS DEBIT,COALESCE(p.COMMENTS,'') AS EXPLANATION,p.PAYMENT_DATE AS DATE,p.ID AS ID
+	$payments_SQL = "SELECT " . $name_col_sql . "'' AS CREDIT,p.AMOUNT AS DEBIT,CONCAT(p.TITLE,' ',COALESCE(p.COMMENTS,'')) AS EXPLANATION,p.PAYMENT_DATE AS DATE,p.ID AS ID
 	FROM accounting_payments p
 	WHERE p.SYEAR='" . UserSyear() . "'
 	AND p.SCHOOL_ID='" . UserSchool() . "'
 	AND p.PAYMENT_DATE BETWEEN '" . $start_date . "'
 	AND '" . $end_date . "'
-	AND STAFF_ID IS NULL";
+	AND STAFF_ID IS NULL" . $where_category_sql;
 
 	$payments_RET = DBGet( $payments_SQL, $extra['functions'] );
 
@@ -264,7 +296,7 @@ if ( isset( $_REQUEST['staff_payroll'], $_REQUEST['student_billing'] ) )
 	$link['add']['html']['STUDENT_NAME'] = '&nbsp;';
 }
 
-$link['add']['html'] = $link['add']['html'] + [
+$link['add']['html'] += [
 	'DEBIT' => '<b>' . Currency( $totals['DEBIT'] ) . '</b>',
 	'CREDIT' => '<b>' . Currency( $totals['CREDIT'] ) . '</b>',
 	'DATE' => '&nbsp;',
@@ -287,4 +319,76 @@ function _makeCurrency( $value, $column )
 	{
 		return Currency( $value );
 	}
+}
+
+/**
+ * Category Select Input
+ *
+ * @since 11.0
+ *
+ * Local function
+ *
+ * @param  string $category Category: lookup in category table.
+ *
+ * @return string Select Category input.
+ */
+function _categorySelect( $category )
+{
+	global $_ROSARIO;
+
+	// Temporary AllowEdit so menu can be viewed in read-only
+	if ( ! AllowEdit() )
+	{
+		$_ROSARIO['allow_edit'] = true;
+
+		$allow_edit_tmp = true;
+	}
+
+	// Build options menu
+	$category_RET = DBGet( "SELECT ID,TITLE,SHORT_NAME,TYPE
+		FROM accounting_categories
+		WHERE SCHOOL_ID='" . UserSchool() . "'
+		ORDER BY TYPE,SORT_ORDER" );
+
+	$category_options = [ '0' => _( 'None' ) ];
+
+	$previous_category = '';
+
+	$category_types = [
+		'common' => _( 'Incomes' ) . ' & ' . _( 'Expenses' ),
+		'incomes' => _( 'Incomes' ),
+		'expenses' => _( 'Expenses' ),
+	];
+
+	foreach ( (array) $category_RET as $category_a )
+	{
+		if ( $previous_category !== $category_a['TYPE'] )
+		{
+			$previous_category = $category_a['TYPE'];
+
+			$category_options[ $previous_category ] = $category_types[ $previous_category ];
+		}
+
+		$category_options[$category_a['ID']] = '&nbsp;&nbsp;&nbsp;' . $category_a['TITLE'];
+	}
+
+	$link = PreparePHP_SELF( [], [ 'category' ] ) . '&category=';
+
+	$menu = SelectInput(
+		$category,
+		'category',
+		'',
+		$category_options,
+		_( 'All' ),
+		'onchange="' . AttrEscape( 'ajaxLink(' . json_encode( $link ) . ' + this.value);' ) . '" autocomplete="off"',
+		false
+	);
+
+	// Temporary AllowEdit removed
+	if ( ! empty( $allow_edit_tmp ) )
+	{
+		$_ROSARIO['allow_edit'] = false;
+	}
+
+	return $menu;
 }